Skip packages we can't find during resolving dependencies
Dub will complain to the user about stuff not being there
and in most causes it is a user that wrote down a wrong dependency in their
dub file and we are forever stuck with it. Changes are they will have
released a fix version so all this falls by the wayside.